requirements.txt
python-pptx==0.6.21
Pillow==9.5.0 